requirements.txt
# Text-to-SQL Pipeline v0.1.2 - Production Requirements
# pip install -r requirements.txt

# ===========================================
# Core
# ===========================================
pydantic>=2.6.0
pydantic-settings>=2.2.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Data Processing
numpy>=1.26.0
polars>=1.0.0
sqlalchemy>=2.0.0

# ===========================================
# LLM - Ollama (Primary Backend)
# ===========================================
ollama>=0.2.0

# ===========================================
# LLM - Transformers (Fallback)
# ===========================================
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.28.0
bitsandbytes>=0.42.0

# ===========================================
# Vector DB
# ===========================================
qdrant-client>=1.7.0

# ===========================================
# Graph DB
# ===========================================
neo4j>=5.0.0

# ===========================================
# Orchestration
# ===========================================
langgraph>=0.0.1

# ===========================================
# Embeddings & Reranking
# ===========================================
sentence-transformers>=2.7.0
cachetools>=5.3.0

# ===========================================
# UI
# ===========================================
streamlit>=1.32.0
pandas>=2.0.0

# ===========================================
# Monitoring
# ===========================================
prometheus-client>=0.19.0
